fix: add-cluster functionality for k8s 1.27 (#715)
## What
In this PR, we are making a simple change to ensure that our CLI
supports adding Kubernetes clusters with version 1.27. The change
involves modifying the way we find the pod owned by a job using the
label.

## Why
These changes are being made to accommodate the changes introduced in
Kubernetes version 1.27. In version 1.27, pods owned by a Job now use
the labels `batch.kubernetes.io/job-name` and
`batch.kubernetes.io/controller-uid`. This is a change from the previous
labels `job-name` and `controller-uid`, which were used for
compatibility. By making this adjustment, we ensure that our CLI is
compatible with Kubernetes 1.27 and can correctly identify pods owned by
a job.

## Notes
Link to the changelog -
https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG/CHANGELOG-1.27.md#api-change-3