feat: update cli to retrieve applications for product release steps (#758)
## What
Adding promotion commits to product release steps if the version of
codefresh is compatible. I've added a fall back in case the cli is
connected to an old version of codefresh (on prem).

## Why
A customer wanted to know the commit created by promotions for each step
of a product release for each application. See Jira ticket for details.

## Notes
Jira: https://codefresh-io.atlassian.net/browse/CR-25720